Separate tests to allow full-unit CI markers
