+++ tests/test_heading_hierarchy.py
@@ -1,9 +1,7 @@
 """Tests for section-header level inference in the PDF/image pipeline."""
 
-from pathlib import Path
 from types import SimpleNamespace
 
-import pytest
 from docling_core.types.doc import (
     BoundingBox,
     CoordOrigin,
@@ -20,22 +18,12 @@ from docling_core.types.doc.page import (
     SegmentedPdfPage,
 )
 
-from docling.backend.docling_parse_backend import DoclingParseDocumentBackend
-from docling.datamodel.accelerator_options import AcceleratorDevice
-from docling.datamodel.base_models import InputFormat
-from docling.datamodel.document import InputDocument
-from docling.datamodel.pipeline_options import (
-    HeadingHierarchyOptions,
-    PdfPipelineOptions,
-)
-from docling.datamodel.settings import DocumentLimits
+from docling.datamodel.pipeline_options import HeadingHierarchyOptions
 from docling.models.stages.heading_hierarchy.heading_hierarchy_model import (
     HeadingHierarchyModel,
     _infer_from_numbering,
     _parse_marker,
 )
-from docling.pipeline.legacy_standard_pdf_pipeline import LegacyStandardPdfPipeline
-from docling.pipeline.standard_pdf_pipeline import StandardPdfPipeline
 
 
 def _levels(texts: list[str], **opts) -> dict[int, int]:
@@ -221,39 +209,3 @@ def test_style_fallback_assigns_levels_by_font_size():
     model.assign_heading_levels(doc, parsed_pages={1: page})
 
     assert [h.level for h in doc.texts] == [1, 2]
-
-
-@pytest.mark.ml_pdf_model
-@pytest.mark.parametrize(
-    "pipeline_cls",
-    [StandardPdfPipeline, LegacyStandardPdfPipeline],
-)
-def test_pdf_pipeline_assigns_heading_levels_from_existing_fixture(
-    pipeline_cls,
-) -> None:
-    pipeline_options = PdfPipelineOptions()
-    pipeline_options.do_ocr = False
-    pipeline_options.do_table_structure = False
-    pipeline_options.generate_parsed_pages = True
-    pipeline_options.accelerator_options.device = AcceleratorDevice.CPU
-    pipeline_options.heading_hierarchy_options = HeadingHierarchyOptions(enabled=True)
-
-    input_document = InputDocument(
-        path_or_stream=Path("tests/data/pdf/2203.01017v2.pdf"),
-        format=InputFormat.PDF,
-        backend=DoclingParseDocumentBackend,
-        limits=DocumentLimits(page_range=(1, 6)),
-    )
-
-    result = pipeline_cls(pipeline_options).execute(
-        input_document, raises_on_error=True
-    )
-    headings = {
-        item.text: item.level
-        for item in result.document.texts
-        if isinstance(item, SectionHeaderItem)
-    }
-
-    assert headings["1. Introduction"] == 1
-    assert headings["4.1. Model architecture."] == 2
-    assert headings["5.1. Implementation Details"] == 2
